Add CPUID availability check for x86-64
The else block basically duplicates the code for the case when USE_64_BIT is set
to False. The only adjustment is to use appropriately sized registers.

Comment: It is unclear to me, why source in the x86-64-specific directory would
need to check the USE_64_BIT flag since x86-64 is inherently 64-bit. The check
and the unneeded code can probably be dropped.